support the other int types

+++ examples/demostruct.go
@@ -2,5 +2,5 @@ package encodedemo
 
 type Demostruct struct {
 	A int64
-	B int64
+	B int32
 }
